requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
branca==0.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
curl_cffi==0.11.4
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
fastapi==0.115.14
fastjsonschema==2.19.1
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
folium==0.20.0
fonttools==4.51.0
fqdn==1.5.1
frozendict==2.4.6
fsspec==2024.3.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
gradio==5.35.0
gradio_client==1.10.4
groovy==0.1.2
grpcio==1.73.1
h11==0.14.0
h5py==3.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.33.2
idna==3.7
ImageHash==4.3.1
imbalanced-learn==0.13.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.5.1
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
keras==3.10.0
kiwisolver==1.4.5
libclang==18.1.1
llvmlite==0.44.0
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
ml_dtypes==0.5.1
mplfinance==0.12.10b0
mpmath==1.3.0
multimethod==1.11.2
multitasking==0.0.11
namex==0.1.0
narwhals==1.46.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
numba==0.61.0
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
overrides==7.7.0
packaging==24.0
pandas==2.3.1
pandas-datareader==0.10.0
pandas-profiling==3.2.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
peewee==3.18.2
phik==0.12.4
pillow==10.3.0
platformdirs==4.2.1
plotly==6.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
puremagic==1.30
pyarrow==15.0.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2024.1
PyWavelets==1.6.0
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
regex==2024.4.28
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
ruff==0.12.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.33.0
sympy==1.14.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
textblob==0.18.0.post0
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.21.2
toml==0.10.2
tomlkit==0.13.3
toolz==0.12.1
torch==2.7.1
torchvision==0.22.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.53.1
typeguard==4.4.4
typer==0.16.0
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.35.0
vaderSentiment==3.3.2
visions==0.8.1
watchdog==4.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.10
wordcloud==1.9.4
wrapt==1.17.2
xyzservices==2025.4.0
ydata-profiling==4.16.1
yfinance==0.2.65
